Add scheduler tests and fix env loading
